+++ api/python/src/DEX/objects/pyParser.cpp
@@ -19,6 +19,7 @@
 #include "LIEF/DEX/File.hpp"
 #include "LIEF/logging.hpp"
 
+#include "pyutils.hpp"
 #include "pyIOStream.hpp"
 
 #include <string>
@@ -47,13 +48,22 @@ void create<Parser>(nb::module_& m) {
     nb::rv_policy::take_ownership);
 
   m.def("parse",
-    [] (nb::object byteio, const std::string& name) -> nb::object {
-      if (auto stream = PyIOStream::from_python(std::move(byteio))) {
-        auto ptr = std::make_unique<PyIOStream>(std::move(*stream));
-        return nb::cast(LIEF::DEX::Parser::parse(stream->content(), name));
+    [] (nb::object obj, const std::string& name) -> nb::object {
+      if (auto path_str = path_to_str(obj)) {
+        return nb::cast(DEX::Parser::parse(std::move(*path_str)));
       }
-      logging::log(logging::LOG_ERR, "Can't create a LIEF stream interface over the provided io");
+
+      if (auto stream = PyIOStream::from_python(obj)) {
+        auto ptr = std::make_unique<PyIOStream>(std::move(*stream));
+        return nb::cast(DEX::Parser::parse(stream->content(), name));
+      }
+
+      logging::log(logging::LOG_ERR,
+                   "LIEF parser interface does not support Python object: " +
+                   type2str(obj));
       return nb::none();
-    }, "io"_a, "name"_a = "", nb::rv_policy::take_ownership);
+    }, "obj"_a, "name"_a = "",
+    "parse(obj: io.IOBase | os.PathLike, name: str = ...) -> Optional[lief.DEX.File]"_p,
+    nb::rv_policy::take_ownership);
 }
 }