+++ api/python/src/OAT/objects/pyParser.cpp
@@ -16,6 +16,7 @@
 #include "OAT/pyOAT.hpp"
 
 #include "pyIOStream.hpp"
+#include "pyutils.hpp"
 
 #include "LIEF/OAT/Parser.hpp"
 #include "LIEF/OAT/Binary.hpp"
@@ -48,13 +49,22 @@ void create<Parser>(nb::module_& m) {
     "raw"_a, nb::rv_policy::take_ownership);
 
   m.def("parse",
-    [] (nb::object byteio) -> nb::object {
-      if (auto stream = PyIOStream::from_python(std::move(byteio))) {
+    [] (nb::object obj) -> nb::object {
+      if (auto path_str = path_to_str(obj)) {
+        return nb::cast(Parser::parse(std::move(*path_str)));
+      }
+
+      if (auto stream = PyIOStream::from_python(obj)) {
         auto ptr = std::make_unique<PyIOStream>(std::move(*stream));
         return nb::cast(Parser::parse(stream->content()));
       }
-      logging::log(logging::LOG_ERR, "Can't create a LIEF stream interface over the provided io");
+      logging::log(logging::LOG_ERR,
+                   "LIEF parser interface does not support Python object: " +
+                   type2str(obj));
       return nb::none();
-    }, "io"_a, nb::rv_policy::take_ownership);
+    },
+    "obj"_a,
+    "parse(obj: io.IOBase | os.PathLike) -> Optional[lief.OAT.Binary]"_p,
+    nb::rv_policy::take_ownership);
 }
 }