+++ api/python/src/PE/objects/pyParser.cpp
@@ -22,7 +22,7 @@
 #include <nanobind/stl/vector.h>
 #include <nanobind/stl/unique_ptr.h>
 
-
+#include "pyutils.hpp"
 #include "pyIOStream.hpp"
 #include "LIEF/logging.hpp"
 
@@ -48,17 +48,22 @@ void create<Parser>(nb::module_& m) {
     nb::rv_policy::take_ownership);
 
   m.def("parse",
-    [] (nb::object byteio, const ParserConfig&) -> nb::object {
-      if (auto stream = PyIOStream::from_python(std::move(byteio))) {
+    [] (nb::object obj, const ParserConfig&) -> nb::object {
+      if (auto path_str = path_to_str(obj)) {
+        return nb::cast(Parser::parse(std::move(*path_str)));
+      }
+      if (auto stream = PyIOStream::from_python(obj)) {
         auto ptr = std::make_unique<PyIOStream>(std::move(*stream));
         return nb::cast(PE::Parser::parse(std::move(ptr)));
       }
-      logging::log(logging::LOG_ERR, "Can't create a LIEF stream interface over the provided io");
+      logging::log(logging::LOG_ERR,
+                   "LIEF parser interface does not support Python object: " +
+                   type2str(obj));
       return nb::none();
     },
-    "Parse the PE binary from the given Python IO interface and return a :class:`lief.PE.Binary` object"_doc,
-    "io"_a, "config"_a = ParserConfig::all(),
-    "parse(io: io.IOBase, config: lief.PE.ParserConfig = ...) -> lief.PE.Binary | None"_p,
+    "Parse the PE binary from the given parameter and return a :class:`lief.PE.Binary` object"_doc,
+    "obj"_a, "config"_a = ParserConfig::all(),
+    "parse(obj: io.IOBase | os.PathLike, config: lief.PE.ParserConfig = ...) -> lief.PE.Binary | None"_p,
     nb::rv_policy::take_ownership);
 }